backend/pyproject.toml
[tool.poetry]
name = "research-portal-backend"
version = "0.1.0"
description = "Backend service for the Python-focused agentic research portal."
authors = ["Your Team <team@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
langchain = "^0.2.0"
langgraph = "^0.0.50"
langsmith = "^0.1.33"
openai = "^1.34.0"
tavily-python = "^0.3.3"
redis = "^5.0.4"
structlog = "^24.1.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
psycopg = { extras = ["binary"], version = "^3.1.19" }
pgvector = "^0.2.4"
alembic = "^1.13.1"
tenacity = "^8.3.0"
langchain-openai = "^0.1.10"
pypdf = "^5.1.0"
markdown = "^3.6"
python-multipart = "^0.0.9"
cryptography = "^43.0.1"
langchain-community = "^0.2.12"
langchain-google-genai = "^1.0.10"
google-generativeai = "^0.7.2"
fpdf2 = "^2.8.1"
opentelemetry-sdk = "^1.26.0"
opentelemetry-exporter-otlp = "^1.26.0"
opentelemetry-instrumentation-fastapi = "^0.47b0"
opentelemetry-instrumentation-requests = "^0.47b0"
prometheus-client = "^0.20.0"
beautifulsoup4 = "^4.12.0"
httpx = "^0.27.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
fakeredis = "^2.23.2"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

